perf: server-side topK heap pushdown for ORDER BY + LIMIT (Tier-3b)
Push the bounded top-K heap from the client down to each group server
when the planner detects ORDER BY + LIMIT. Each group runs the heap
locally on its decoded DataRows and ships only the K winners over the
wire. The client receives at most K*Ngroups candidates and does a
trivial k-way merge.

Changes:
- protocol/messages/kv: WireTopKSpec + WireSortSpec types, ScanRequest.topK
  field, ScanFlagHasTopK (0x100). Promote ScanRequest.flags from u8 to u16.
- protocol/server: scan handler detects req.topK, accumulates projected
  rows in a captured seq[seq[string]] instead of sending frames, runs
  TopKHeap after scan, ships ONE final frame with K winners.
- sql/planner: scTopK on poScan, obServerTopK on poOrderBy. Set when
  obOptimization == oboTopK and limit > 0 and sort specs are simple
  column refs.
- sql/executor: poOrderBy oboTopK branch skips client-side heap when
  op.obServerTopK is true. Stream iterator consumes K candidates directly.
- client/fractio_client: threads carry serialized WireTopKSpec via
  SetupArg ('limit|columnIndex,descending|...').

Benchmarks (10K rows, 3-node cluster):
  T5: ORDER BY name LIMIT 5         70ms -> 0.32ms (220x)
  T6: ORDER BY name DESC LIMIT 5    70ms -> 0.32ms (220x)
  T7: ORDER BY name (full sort)     92ms -> 81ms   (1.13x)
  T8: 2-col proj + sort LIMIT 5     ~70ms -> 0.31ms
  T9: 1-col proj + sort LIMIT 5     ~70ms -> 0.32ms

Wire format change: ScanRequest.flags is now uint16. No backward
compatibility (per user request).

Tests: 115 unit tests pass. Test fixtures updated for u8 -> u16 flag
constants and readUint8 -> readUint16BE for flag fields.

Pre-existing bug found during benchmarking (not in this commit):
  SELECT id, name FROM t WHERE value > 5000 ORDER BY name LIMIT 5
  returns 0 rows because scColumns excludes filter-referenced columns.
  Tracked separately.

+++ tests/unit/protocol/messages/test_kv.nim
@@ -772,11 +772,11 @@ suite "KV Constants":
     check BatchStatusAllFailed == 0x02'u8
 
   test "Scan flag values":
-    check ScanFlagIncludeTimestamp == 0x01'u8
-    check ScanFlagIncludeVersion == 0x02'u8
-    check ScanFlagKeysOnly == 0x04'u8
-    check ScanFlagReverse == 0x08'u8
-    check ScanFlagGroupRouted == 0x10'u8
+    check ScanFlagIncludeTimestamp == 0x01'u16
+    check ScanFlagIncludeVersion == 0x02'u16
+    check ScanFlagKeysOnly == 0x04'u16
+    check ScanFlagReverse == 0x08'u16
+    check ScanFlagGroupRouted == 0x10'u16
 
   test "Scan response flag values":
     check ScanRespFlagHasMore == 0x01'u8
